fix(latency): stop bottoming out at 0ms across the dashboard
StyloBot detection runs in microseconds-to-low-milliseconds, but every
processing-time display rounded sub-ms values down to "0ms" -- the operator
couldn't tell a 320µs detector apart from a no-op one ("we don't even SHOW
how long it took"). Cross-cutting unit-bucketed formatter end to end:

  * BlackboardOrchestrator.cs:1026: dropped the `(long)` cast on
    Stopwatch.GetElapsedTime(...).TotalMilliseconds. The HandleDetectorFailure
    signature already takes double; the cast was discarding every sub-ms
    detector's measurement before it ever left the orchestrator.

  * AnalyticsFormatting.FormatMs (the canonical server-side formatter): now
    buckets the value -- "-" for non-positive, "<1µs" for sub-microsecond,
    "{N}µs" up to 1ms, "{N.NN}ms" up to 10ms, "{N}ms" up to 1s, "{N.N}s"
    above. Pinned by 15 inline-data tests covering each bucket boundary.

  * _SignatureDetail.cshtml: footer "Processing:" and endpoint stats columns
    route through FormatMs. The Processing Time chart's JS formatter mirrors
    the server-side bucketing so the Y-axis ticks and the on-hover tooltip
    agree (and a 320µs request actually shows as "320µs").

  * BotDetectionHeader / BotDetectionDetails Compact widgets: Alpine x-text
    expressions inlined with the same JS bucketing logic so the live YOUR
    DETECTION badge stops claiming "0ms".

+++ src/Mostlylucid.BotDetection/Orchestration/BlackboardOrchestrator.cs
@@ -1023,7 +1023,7 @@ public class BlackboardOrchestrator : IDetectionOrchestrator
                 contributions = await detector.ContributeAsync(state, cancellationToken);
             }
 
-            var elapsedMs = (long)Stopwatch.GetElapsedTime(startTs).TotalMilliseconds;
+            var elapsedMs = Stopwatch.GetElapsedTime(startTs).TotalMilliseconds;
 
             foreach (var contribution in contributions)
             {
@@ -1047,12 +1047,12 @@ public class BlackboardOrchestrator : IDetectionOrchestrator
         }
         catch (OperationCanceledException) when (!cancellationToken.IsCancellationRequested)
         {
-            var elapsedMs = (long)Stopwatch.GetElapsedTime(startTs).TotalMilliseconds;
+            var elapsedMs = Stopwatch.GetElapsedTime(startTs).TotalMilliseconds;
             HandleDetectorFailure(detector, aggregator, failedDetectors, "Timeout", elapsedMs);
         }
         catch (Exception ex)
         {
-            var elapsedMs = (long)Stopwatch.GetElapsedTime(startTs).TotalMilliseconds;
+            var elapsedMs = Stopwatch.GetElapsedTime(startTs).TotalMilliseconds;
             HandleDetectorFailure(detector, aggregator, failedDetectors, ex.Message, elapsedMs);
 
             _logger.LogWarning(ex,
